+++ src/main/java/study/step3/Car.java
@@ -1,4 +1,4 @@
-package study;
+package study.step3;
 
 public interface Car {
     boolean isMoved();
