+++ src/main/java/study/step3/CarRacing.java
@@ -1,4 +1,4 @@
-package study;
+package study.step3;
 
 import java.util.HashSet;
 import java.util.Set;
