+++ src/main/java/study/step3/RacingInfoProvider.java
@@ -1,4 +1,4 @@
-package study;
+package study.step3;
 
 import java.util.Set;
