+++ src/test/java/study/step3/RealCarTest.java
@@ -1,7 +1,8 @@
-package study;
+package study.step3;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import study.step3.RealCar;
 
 import static org.assertj.core.api.Assertions.assertThat;
