+++ src/test/java/study/step3/ResultViewTest.java
@@ -1,4 +1,4 @@
-package study;
+package study.step3;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
